+++ src/NLSS.jl
@@ -1,6 +1,7 @@
 module NLSS
 
 include("Simulation.jl")
+include("Utilities.jl")
 include("Plotter.jl")
 
 end #module
