+++ src/SimSolvers.jl
@@ -1,4 +1,4 @@
-export solve!
+
 """
     solve!(sim::Simulation)
 
@@ -55,6 +55,9 @@ function solve!(sim::Sim)
     return nothing
 end #solve
 
+#######################################################################################
+# Algorithms
+#######################################################################################
 """
     T2(ψ, ω, dx, F)
